fixed linting in LegendBar.tsx
